# swagger.yaml
openapi: 3.0.3
info:
  title: Event Booking System API
  description: |
    A Mini Event Management System where users can browse events and book tickets.
    Built with Node.js (Express), MySQL, and documented via OpenAPI 3.0.
  version: 1.0.0
  contact:
    name: Rahul Pal
    email: rahulsinghpal155@gmail.com

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Events
    description: Event management endpoints
  - name: Bookings
    description: Ticket booking endpoints
  - name: Users
    description: User-related endpoints
  - name: Attendance
    description: Event attendance / check-in endpoints

paths:
  # ──────────────────────────────────────────────────────────────
  # GET /events
  # ──────────────────────────────────────────────────────────────
  /events:
    get:
      tags:
        - Events
      summary: List all upcoming events
      description: Returns all events whose date is >= now, ordered by date ascending.
      operationId: getAllEvents
      responses:
        '200':
          description: A list of upcoming events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
              example:
                success: true
                count: 2
                data:
                  - id: 1
                    title: Tech Summit 2026
                    description: Annual technology conference
                    date: '2026-06-15T09:00:00.000Z'
                    total_capacity: 200
                    remaining_tickets: 198
                  - id: 2
                    title: Node.js Workshop
                    description: Hands-on Node.js for beginners
                    date: '2026-07-20T10:00:00.000Z'
                    total_capacity: 50
                    remaining_tickets: 50
        '500':
          $ref: '#/components/responses/InternalServerError'

    # ──────────────────────────────────────────────────────────────
    # POST /events
    # ──────────────────────────────────────────────────────────────
    post:
      tags:
        - Events
      summary: Create a new event
      description: Creates a new event. `remaining_tickets` is automatically set equal to `total_capacity`.
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
            example:
              title: React Conference 2026
              description: Deep dive into React 19 and server components
              date: '2026-09-10T09:00:00'
              total_capacity: 150
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleEventResponse'
              example:
                success: true
                message: Event created successfully
                data:
                  id: 4
                  title: React Conference 2026
                  description: Deep dive into React 19 and server components
                  date: '2026-09-10T09:00:00.000Z'
                  total_capacity: 150
                  remaining_tickets: 150
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ──────────────────────────────────────────────────────────────
  # POST /bookings
  # ──────────────────────────────────────────────────────────────
  /bookings:
    post:
      tags:
        - Bookings
      summary: Book a ticket for a user
      description: |
        Books a ticket for a user on an event.
        - Uses a database transaction with `SELECT FOR UPDATE` to prevent race conditions on `remaining_tickets`.
        - Generates a unique UUID `booking_code` returned to the user.
        - Returns 409 if no tickets remain or the user has already booked this event.
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
            example:
              user_id: 1
              event_id: 2
      responses:
        '201':
          description: Ticket booked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingCreatedResponse'
              example:
                success: true
                message: Ticket booked successfully
                data:
                  booking_id: 7
                  user_id: 1
                  event_id: 2
                  event_title: Node.js Workshop
                  booking_code: 550e8400-e29b-41d4-a716-446655440000
                  booking_date: '2026-03-21T13:45:00.000Z'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict — no tickets remaining or duplicate booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                noTickets:
                  summary: No tickets remaining
                  value:
                    success: false
                    message: No tickets remaining for this event
                duplicate:
                  summary: Duplicate booking
                  value:
                    success: false
                    message: User has already booked this event
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ──────────────────────────────────────────────────────────────
  # GET /users/:id/bookings
  # ──────────────────────────────────────────────────────────────
  /users/{id}/bookings:
    get:
      tags:
        - Users
      summary: Retrieve all bookings for a user
      description: Returns every booking made by the specified user, ordered by booking date descending.
      operationId: getUserBookings
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: List of bookings for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookingsResponse'
              example:
                success: true
                user:
                  id: 1
                  name: Alice Johnson
                  email: alice@example.com
                count: 1
                data:
                  - booking_id: 7
                    booking_code: 550e8400-e29b-41d4-a716-446655440000
                    booking_date: '2026-03-21T13:45:00.000Z'
                    event_id: 2
                    event_title: Node.js Workshop
                    event_description: Hands-on Node.js for beginners
                    event_date: '2026-07-20T10:00:00.000Z'
                    total_capacity: 50
                    remaining_tickets: 49
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ──────────────────────────────────────────────────────────────
  # POST /events/:id/attendance
  # ──────────────────────────────────────────────────────────────
  /events/{id}/attendance:
    post:
      tags:
        - Attendance
      summary: Record event attendance via booking code
      description: |
        Validates a `booking_code` for the given event and records the attendee's entry.
        Returns the total number of tickets booked for the event.
        Returns 409 if the code has already been used for entry.
      operationId: recordAttendance
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceRequest'
            example:
              booking_code: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Attendance recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
              example:
                success: true
                message: Attendance recorded successfully
                data:
                  event_id: 2
                  event_title: Node.js Workshop
                  attendee:
                    user_id: 1
                    name: Alice Johnson
                    email: alice@example.com
                  entry_time: '2026-07-20T10:05:00.000Z'
                  total_bookings_for_event: 12
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Ticket already used for entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: This ticket has already been used for entry
        '500':
          $ref: '#/components/responses/InternalServerError'

# ══════════════════════════════════════════════════════════════
# Components
# ══════════════════════════════════════════════════════════════
components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: Unique user ID
      schema:
        type: integer
        minimum: 1
      example: 1

    EventId:
      name: id
      in: path
      required: true
      description: Unique event ID
      schema:
        type: integer
        minimum: 1
      example: 2

  schemas:
    # ── Request bodies ─────────────────────────────────────────
    CreateEventRequest:
      type: object
      required:
        - title
        - date
        - total_capacity
      properties:
        title:
          type: string
          maxLength: 255
          example: Tech Summit 2026
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: Annual technology conference
        date:
          type: string
          format: date-time
          description: Must be a future date/time (ISO 8601)
          example: '2026-09-10T09:00:00'
        total_capacity:
          type: integer
          minimum: 1
          example: 150

    CreateBookingRequest:
      type: object
      required:
        - user_id
        - event_id
      properties:
        user_id:
          type: integer
          minimum: 1
          example: 1
        event_id:
          type: integer
          minimum: 1
          example: 2

    AttendanceRequest:
      type: object
      required:
        - booking_code
      properties:
        booking_code:
          type: string
          description: UUID v4 booking code provided at time of booking
          example: 550e8400-e29b-41d4-a716-446655440000

    # ── Domain models ──────────────────────────────────────────
    Event:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Tech Summit 2026
        description:
          type: string
          nullable: true
          example: Annual technology conference
        date:
          type: string
          format: date-time
          example: '2026-06-15T09:00:00.000Z'
        total_capacity:
          type: integer
          example: 200
        remaining_tickets:
          type: integer
          example: 198

    BookingDetail:
      type: object
      properties:
        booking_id:
          type: integer
          example: 7
        booking_code:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        booking_date:
          type: string
          format: date-time
          example: '2026-03-21T13:45:00.000Z'
        event_id:
          type: integer
          example: 2
        event_title:
          type: string
          example: Node.js Workshop
        event_description:
          type: string
          nullable: true
          example: Hands-on Node.js for beginners
        event_date:
          type: string
          format: date-time
          example: '2026-07-20T10:00:00.000Z'
        total_capacity:
          type: integer
          example: 50
        remaining_tickets:
          type: integer
          example: 49

    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Alice Johnson
        email:
          type: string
          format: email
          example: alice@example.com

    # ── Response wrappers ──────────────────────────────────────
    EventListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          example: 2
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'

    SingleEventResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Event created successfully
        data:
          $ref: '#/components/schemas/Event'

    BookingCreatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Ticket booked successfully
        data:
          type: object
          properties:
            booking_id:
              type: integer
              example: 7
            user_id:
              type: integer
              example: 1
            event_id:
              type: integer
              example: 2
            event_title:
              type: string
              example: Node.js Workshop
            booking_code:
              type: string
              example: 550e8400-e29b-41d4-a716-446655440000
            booking_date:
              type: string
              format: date-time
              example: '2026-03-21T13:45:00.000Z'

    UserBookingsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        count:
          type: integer
          example: 1
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookingDetail'

    AttendanceResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Attendance recorded successfully
        data:
          type: object
          properties:
            event_id:
              type: integer
              example: 2
            event_title:
              type: string
              example: Node.js Workshop
            attendee:
              $ref: '#/components/schemas/User'
            entry_time:
              type: string
              format: date-time
              example: '2026-07-20T10:05:00.000Z'
            total_bookings_for_event:
              type: integer
              example: 12

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred

    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: total_capacity
              message:
                type: string
                example: Total capacity must be a positive integer

  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'

    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Event not found

    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Internal Server Error